To-Do App
This is a simple To-Do App built with HTML, CSS, and JavaScript to help you manage your tasks. The app has a clean and simple design with alternating background colors for each task. It also includes hover effects to make the app more interactive and user-friendly. This project is great for beginners to practice basic web development skills like creating interactive web pages, handling user input, and using JavaScript to manipulate the page.